changed the added noise in detectability plots
Was fixed at 0.1 radm^-2 and was supposed to be due to instrument noise, but didn't make physical sense.  Now comes from noise far from filaments in standard deviation of RM sightlines outside all halos, with value 0.15 radm^-2, making filaments a bit harder to detect.  Made sure to propagate this so that the filament width was also correctly measured.  Also added explanation for changing the number of sources per square degree.